requirements.txt
pdfminer.six == 20231228
python-dateutil ==2.9.0.post0